Pull config classes into separate file, Add server port to config file, Expose config to app via global
