+++ server/config/settings.yaml
@@ -1,4 +1,5 @@
 bridge:
+  port: 3001
   enable_auth: false
   enable_https: false
 ecobee: {}
